% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpsy.R
\name{loadexp}
\alias{loadexp}
\title{Load Experimental Files from Server}
\usage{
loadexp(expfolder)
}
\arguments{
\item{expfolder}{The data folder of the experiment to be loaded}
}
\value{
Dataframe containing all trials of the combines result files
}
\description{
Read all results files (.csv) from the server and write to dataframe
}
\examples{
temp1 <- C_to_F(22);
temp2 <- C_to_F( c(-2, 12, 23) );
}
