% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpsy.R
\name{splitresponse}
\alias{splitresponse}
\title{Write responses into variables.}
\usage{
splitresponse(data)
}
\value{
Generates new variables (columns) named after the question and storing the given responses.
}
\description{
Write all answers from the "response"-column generated by a survey-plugin into individual variables.
}
\examples{
demo <- myexp \%>\%
  filter(trial_part="demographics") \%>\%
  splitresponse()
}
