% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpsy.R
\name{loadexp}
\alias{loadexp}
\title{Load Experimental Files from Server}
\usage{
loadexp(exp, subdirs = TRUE)
}
\arguments{
\item{exp}{String. The name of the data folder of the experiment to be loaded.}

\item{subdirs}{Logical. Should subdirectories be considered?}
}
\value{
Dataframe containing all trials of the combined result files.
}
\description{
Read all results files (.csv) from the server and write to dataframe
}
\examples{
myexp <- loadexp("myexp")
}
